Add files from PelicanHPC website
The files in the "Original_PelicanHPC_Files" folder were retrvived in
approximately 2017 from the currently defunct website pelicanhpc.org,
and are the necessary components to build the iso image. The scripts in
the compressed files will need to be updated before they will work, but
will be the starting point for this new project.