Add net-tools to package list, modify scripts to /usr/sbin
Added net-tools to list of packages to include, on both .gnome and .xfce
variants.  This package is not included by default anymore, it is the
source for the ifconfig utility which is used in many of the
scripts.

Additionally, the location of the ifconfig binary has moved to
/usr/sbin, so modified the scripts to point to this, to eliminate any
troubles.

Sometime, will want to modify the scripts to use the iptools utility,
but will push that off till later