Creating a "No GUI" option, in addtion to gnome and xfce
Couple  of reasons.  Am troubleshooting an issuse with the nodes being
unresponsive and slow, think it has something to do with the NFSmounts.
Am seeing a lot of audio related file connections when runing "lsof -N
/home"  (pipwire, pulseaudio, etc).  Thinking that removing the GUI
(xfce or gnome) packages will keept these packages from being installed

Additionally, think I will mount the iso file on a ramdisk (for the
permanant server option anyway), so a minimal iso filesize will be
helpful for limiting the amount of ram use for that.  Thinking that
having the iso in ram will speed things up.  Will see