Subustitute ip commands for ifconfg commands in pelcian_boot_setup
ifconfing is deprecated. Also, the format has apparently changed over
the years, so that getting the ip address to set as a variable in the
script was failing.  Replaced lines in this scripit with ip.

Also, am using the entire ip address as part of the node hostname,
rather than just the last octet. (at least I think that's what was
happening)