Modified the network setup, using new commands from modern linux
Changed from using "ifconfig" to "ip" because apparently thats how
things are done these days

If a mac address was given as the $CLUSTER_NETWORK in the
"pelican_config" file (located in the home folder), the script would
convert it to a interface name.  Had to change this logic, as interface
names are not called "eth_" anymore.